requirements.txt
flask
tensorflow
gunicorn
azure.storage.blob
